Complete TakerStateMachine refactor
